requirements.txt
paddlepaddle-gpu
paddlespeech
paddleaudio
paddle2onnx